Emit DebugInfo for entry point parameters
This commit is to emit the debug-info for the entry point parameters.

The problem was when the entry point parameter contains varying-input or
system built-in semtantics, they are legalized to global uniform
variables and the entry point parameter itself ends up having no member
variables at all. That caused the entry point variable to be completely
removed and there was no debug information emitted for it.

This commit emits `DebugLocalVariable` for the entry point paramter.
And its members are properly:
  1. emited as `DebugLocalVariable` and
  2. mapped to the corresponding global variable with `DebugValue`.

In the test code, you will find
```
struct PSIn
{
   float4 pos      : SV_Position;
};

float4 main(PSIn input) : SV_TARGET { ... }
```
With this commit, slang emits the following debug-info:
```
%input = ... DebugLocalVariable "input" ...
%input_pos = ... DebugLocalVariable "input.pos" ...
DebugValue %input_pos %gl_FragCoord
```