Add Optional<T> to Optional<U> implicit conversion
Allow `Optional<Foo>` to implicitly convert to `Optional<IFoo>` when `Foo`
implements `IFoo`, supporting both concrete types and associated types.

The handling of implicit convert happens in `_coerce()` and we are at
AST processing stage not IR. That's why a new type `ConvertOptionalExpr`
is added. It will be lowered to IR when we visit the node.