Add "ignore" option to SLANG_ASSERT environment variable
When the environment variable `SLANG_ASSERT` is set to "ignore" on
Windows, any assertion failures will be ignored. This behavior will be
same to how Release build works and it can help us to reproduce issues
that shows different behavior between Release and Debug builds.

A new function `handleAssert()` had to be introduced, because the
pre-existing `handleSignal()` has [[noreturn]] which cannot work if we
want to just return and resume the normal flow.